Added service to voting and user and change logic, put the logic business in service

+++ src/main/java/com/br/repository/VotingRepository.java
@@ -9,6 +9,6 @@ import org.springframework.data.repository.query.Param;
 
 import com.br.model.Voting;
 
-public interface VotingRepository extends CrudRepository<Voting, Long> {
+public interface VotingRepository extends CrudRepository<Voting, Integer> {
 	Optional<Voting> findByIdUser(int idUser);
 }
